src/requirements/base.txt
django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
django-autoslug==1.9.9
django-countries==7.5.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.24
drf-yasg==1.21.7
Pillow==10.1.0
argon2-cffi==23.1.0
pytz==2023.3.post1
django-taggit==5.0.1
redis==5.0.1
celery==5.3.5
flower==2.0.1
django-celery-email==3.0.0
sql-inspect>=1.0.0
django-allauth==0.60.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2